# packages/backend/src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Notepad
host: "notepad.ahmet.se:3001"
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /note:
    x-swagger-router-controller: note
    get:
      summary: Get notes
      operationId: getNotes
      security:
        - Bearer: []
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/NotesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      summary: Update an existing pet
      operationId: deleteNote
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          description: Note to be removed
          required: true
          schema:
            type: object
            required:
              - id
            properties:
              id:
                type: string
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Note  not found
        "405":
          description: Validation exception
    post:
      summary: Add a new note
      operationId: addNote
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          description: Pet object that needs to be added to the store
          required: true
          schema:
            $ref: "#/definitions/Note"
      responses:
        "405":
          description: Invalid input
    put:
      summary: Update an existing pet
      operationId: updateNote
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          description: Pet object that needs to be added to the store
          required: true
          schema:
            $ref: "#/definitions/Note"
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Pet not found
        "405":
          description: Validation exception
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >
      For accessing the API a valid JWT token must be passed in all the queries
      in

      the 'Authorization' header.


      The following syntax must be used in the 'Authorization' header :
          Bearer xxxxxx.yyyyyyy.zzzzzz
definitions:
  Note:
    type: object
    required:
      - message
    properties:
      message:
        type: string
  NotesResponse:
    type: array
    items:
      $ref: "#/definitions/Note"
  Response:
    type: object
    properties:
      code:
        type: integer
        format: int32
      type:
        type: string
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string